--- FILE201/Database_Connection/.env
@@ -1,5 +1,0 @@
-#database connection
-DB_DATABASE=FILE201
-DB_HOST=localhost
-DB_PASSWORD=
-DB_USER=root
