--- FILE201/Database_Connection/DBConnection.py
@@ -1,57 +1,0 @@
-import sys
-import os
-import mysql.connector
-from mysql.connector import Error
-import logging
-from dotenv import load_dotenv
-
-def resource_path(relative_path):
-    try:
-        base_path = sys._MEIPASS2
-    except Exception:
-        base_path = os.path.abspath(".")
-    return os.path.join(base_path, relative_path)
-
-# Load environment variables from .env file if it exists
-dotenv_path = resource_path("FILE201\\Database_Connection\\.env")
-load_dotenv(dotenv_path)
-
-# Configure logging
-#logging.basicConfig(level=logging.INFO, filename='app.log', filemode='a', format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-#logger = logging.getLogger(__name__)
-
-def create_connection():
-    try:
-        #host = os.getenv('DB_HOST', '192.168.1.2')
-        host = os.getenv('DB_HOST', '127.0.0.1')
-        database = os.getenv('DB_DATABASE')
-        user = os.getenv('DB_USER')
-        password = os.getenv('DB_PASSWORD')
-        port = int(os.getenv('DB_PORT', 3306))  # Default MySQL port is 3306
-
-        if not user or not database:
-            raise ValueError("Database user and name must be provided")
-
-        connection = mysql.connector.connect(
-            host=host,
-            database=database,
-            user=user,
-            password=password,
-            port=port
-        )
-        if connection.is_connected():
-            #logger.info("Connected to MySQL database")
-            print("Connection to MySQL database was successful")
-            return connection
-        else:
-            #logger.info("Failed to connect to MySQL database")
-            print("Failed to connect to MySQL database")
-            return None
-    except Error as e:
-        #logger.exception("Error while connecting to MySQL: %s", e)
-        print("Error while connecting to MySQL:", e)
-        return None
-    except ValueError as ve:
-        #logger.error("Configuration error: %s", ve)
-        print("Configuration error:", ve)
-        return None
