+++ FILE201/Employee_List/employeeList.ui
@@ -266,7 +266,7 @@ background-color: &quot;#DCE5FE&quot;;
       <string/>
      </property>
      <property name="pixmap">
-      <pixmap>../Resources/Icons/search.svg</pixmap>
+      <pixmap>../../MainFrame/Resources/Icons/search.svg</pixmap>
      </property>
      <property name="scaledContents">
       <bool>false</bool>
@@ -304,7 +304,7 @@ background-color: &quot;#DCE5FE&quot;;
     </property>
     <property name="icon">
      <iconset>
-      <normaloff>../Resources/Icons/user-plus.svg</normaloff>../Resources/Icons/user-plus.svg</iconset>
+      <normaloff>../../MainFrame/Resources/Icons/user-plus.svg</normaloff>../../MainFrame/Resources/Icons/user-plus.svg</iconset>
     </property>
     <property name="iconSize">
      <size>
@@ -341,7 +341,7 @@ background-color: &quot;#DCE5FE&quot;;
     </property>
     <property name="icon">
      <iconset>
-      <normaloff>../Resources/Icons/export.svg</normaloff>../Resources/Icons/export.svg</iconset>
+      <normaloff>../../MainFrame/Resources/Icons/export.svg</normaloff>../../MainFrame/Resources/Icons/export.svg</iconset>
     </property>
     <property name="iconSize">
      <size>
