+++ FILE201/file201_Function/excelExport.py
@@ -1,14 +1,13 @@
-import logging
 import pandas as pd
 
 from mysql.connector import Error
-from FILE201.Database_Connection.DBConnection import create_connection
+from MainFrame.Database_Connection.DBConnection import create_connection
 
 #logger = logging.getLogger(__name__)
 
 def fetch_personal_information():
     try:
-        connection = create_connection()
+        connection = create_connection('FILE201')
         if connection is None:
             print("Error: Could not establish database connection.")
             return None, None
