+++ MainFrame/Database_Connection/.env
@@ -1,0 +1,11 @@
+# FILE201 database connection
+DB_DATABASE_FILE201=FILE201
+DB_HOST_FILE201=localhost
+DB_PASSWORD_FILE201=
+DB_USER_FILE201=root
+
+# TIMEKEEPING database connection
+DB_DATABASE_TIMEKEEPING=TIMEKEEPING
+DB_HOST_TIMEKEEPING=localhost
+DB_PASSWORD_TIMEKEEPING=
+DB_USER_TIMEKEEPING=root
