+++ TimeKeeping/timeSheet/TimeSheet.ui
@@ -252,7 +252,7 @@
      <x>-10</x>
      <y>109</y>
      <width>1711</width>
-     <height>641</height>
+     <height>555</height>
     </rect>
    </property>
    <property name="sizePolicy">
